The bitcoin port is added by default to the rpchost.
A snapshot method is added to be able to compare the config file at
launch vs when changes are detected.

A file_watcher object is added that monitor the file for changes and
invokes methods to respond when changes are detected.

Methods are added to respond to changes in lnd and bitcoin config file
changes detected.

A property has been added to determine whether a restart is required to
the lnd node.